requirements.txt
Django
mysqlclient
Pillow
django-crispy-forms
django-ckeditor
djangorestframework
markdown
django-filter
pytz